refactor(gridsat): rename netcdf_image module to image
