+++ src/goesdl/gridsat/__init__.py
@@ -10,7 +10,7 @@ from .dataset_info import GSDatasetInfo
 from .locator_b1 import GridSatProductLocatorB1
 from .locator_gc import GridSatProductLocatorGC
 from .netcdf_geodetic import GSLatLonGrid
-from .netcdf_image import GSImage
+from .image import GSImage
 from .netcdf_info import (
     GSGeospatialInfo,
     GSPlatformInfo,
