+++ src/goesdl/gridsat/image.py
