+++ src/goesdl/gridsat/utility.py
@@ -2,7 +2,7 @@ from netCDF4 import Dataset
 
 from ..geodesy import RectangularRegion
 from ..gridsat.netcdf_geodetic import GSLatLonGrid
-from ..gridsat.netcdf_image import GSImage
+from .image import GSImage
 from .time import GSCoverageTime
 from .dataset_info import GSDatasetInfo
